feat : Collinear한 경우 Segment로 관리.
problem : normalized 공간에서 Delaunay Triangle에 속하지 않는 점들에 대해서 보정이 필요함.

+++ Framework/Model/Mesh/CSkeletal.cpp
@@ -47,7 +47,7 @@ const CBone* CSkeletal::FindBone(const string& InBoneName) const
 
 const CBone * CSkeletal::FindBone(int Index) const
 {
-	if (Index < 0 || Index >= Bones.size())
+	if (Index < 0 || Index >= static_cast<int>(Bones.size()))
 		return nullptr;
 	return Bones[Index];
 }
